+++ jslint.mjs
@@ -163,7 +163,7 @@ let jslint_charset_ascii = (
     + "@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\\]^_"
     + "`abcdefghijklmnopqrstuvwxyz{|}~\u007f"
 );
-let jslint_edition = "v2025.11.1-beta";
+let jslint_edition = "v2025.12.1-beta";
 let jslint_export;                      // The jslint object to be exported.
 let jslint_fudge = 1;                   // Fudge starting line and starting
                                         // ... column to 1.
@@ -4674,7 +4674,13 @@ function jslint_phase3_parse(state) {
 
             test_cause("free");
             the_paren.free = true;
-            if (the_argument.wrapped === true) {
+            if (
+                the_argument.wrapped === true
+
+// PR-483 - Allow parenthesis after ellipsis inside a function call.
+
+                && the_argument.ellipsis !== true
+            ) {
 
 // test_cause:
 // ["aa((0))", "infix_lparen", "unexpected_a", "(", 3]